% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{fit_intercepts}
\alias{fit_intercepts}
\title{Fit an intercept-only binary choice model}
\usage{
fit_intercepts(data, prior = NULL, bayes = FALSE, ...)
}
\arguments{
\item{data}{A list of data values:
\itemize{
\item \code{Y} an \eqn{N \times J} matrix of \eqn{\pm 1} or \code{NA}.
\item \code{N}, \code{J} the dimensions of \code{Y}.
\item \code{D} the dimensions of political conflict being modeled.
}}

\item{prior}{A list of priors for the parameters ab, gamma:
\itemize{
\item \code{sigma.inv.ab} a \eqn{(D+1)\times (D+1)} positive definite matrix corresponding to the inverse of the prior covariance matrix for \eqn{(\alpha, \beta)}.
\item \code{mu.ab} a \eqn{(D+1)} vector corresponding to the prior mean of \eqn{(\alpha, \beta)}.
\item \code{sigma.inv.gamma} a \eqn{D \times D} definite matrix corresponding to the inverse of the prior covariance matrix for \eqn{\gamma}.
\item \code{mu.gamma} a \eqn{D} vector corresponding to the prior mean of \eqn{\gamma}.
}}

\item{bayes}{Logical, set to true if the MAP estimate is desired
instead of the MLE.}
}
\value{
A list only containing intercepts \code{alpha}.
}
\description{
Do not call directly! Used by multiscale
}
