% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{make_starts}
\alias{make_starts}
\title{Make Parameter Initialization Values}
\usage{
make_starts(data)
}
\arguments{
\item{data}{A list of data values:
\itemize{
\item \code{Y} an \eqn{N \times J} matrix of \eqn{\pm 1} or \code{NA}.
\item \code{N}, \code{J} the dimensions of \code{Y}.
\item \code{D} the dimensions of political conflict being modeled.
}}
}
\value{
A list of initialized parameter values for \code{alpha},
    \code{beta}, and \code{gamma}
}
\description{
Make Parameter Initialization Values
}
