% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_main.R
\name{multiscale}
\alias{multiscale}
\title{Fit a zero- or multi-dimensional spatial voting model.}
\usage{
multiscale(method = "sparse", prior, data, init, max.iter = 250,
  tol = 1e-04, verbose = TRUE)
}
\arguments{
\item{method}{One of \code{c("sparse", "intercepts")},
with \code{"sparse"} as the default. If \code{"intercepts"},
estimate an intercept-only model. Otherwise, estimate a model
with positive dimension.}

\item{prior}{A list of priors for the parameters ab, gamma:
\itemize{
\item \code{sigma.inv.ab} a \eqn{(D+1)\times (D+1)} positive definite matrix corresponding to the inverse of the prior covariance matrix for \eqn{(\alpha, \beta)}.
\item \code{mu.ab} a \eqn{(D+1)} vector corresponding to the prior mean of \eqn{(\alpha, \beta)}.
\item \code{sigma.inv.gamma} a \eqn{D \times D} definite matrix corresponding to the inverse of the prior covariance matrix for \eqn{\gamma}.
\item \code{mu.gamma} a \eqn{D} vector corresponding to the prior mean of \eqn{\gamma}.
}}

\item{data}{A list of data values:
\itemize{
\item \code{Y} an \eqn{N \times J} matrix of \eqn{\pm 1} or \code{NA}.
\item \code{N}, \code{J} the dimensions of \code{Y}.
\item \code{D} the dimensions of political conflict being modeled.
}}

\item{init}{A list of initialization values for \code{alpha}, \code{beta}, \code{gamma}.}

\item{max.iter}{The maximum number of iterations.}

\item{tol}{The algorithm stops after the current iteration and the previous iteration parameters have correlation 1 - \code{tol}.}

\item{verbose}{print useful warnings and updates on algorithm progress.}
}
\value{
A list of estimated parameters \code{alpha}, \code{beta}, \code{gamma}
}
\description{
The sparse (as opposed to dense) algorithm ignores the latent
variables for missing vote observations when updating the global
parameters alpha, beta, gamma. It is robust in the presence of many
missing votes.
}
\examples{
\dontrun{
data(s109) # from pscl package
data(s109)
Y <- s109$votes  ## WARNING: data must be +/- 1, or NA
Y[Y \%in\% 1:3] <- 1
Y[Y \%in\% 4:6] <- -1
Y[Y \%in\% c(0, 7:9)] <- NA
data <- list(Y = Y, N = dim(Y)[1], J = dim(Y)[2], D = 2)
prior <- make_prior(data)
init <- make_starts(data)
lout <- multiscale(method = "sparse", prior = prior, data = data, init = init)
cor(lout$gamma)
}
}
