% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense_main.R
\name{multiscale_dense}
\alias{multiscale_dense}
\title{Multidimensional scaling WITHOUT large amounts of missing data.}
\usage{
multiscale_dense(prior, data, init, max.iter = 100, tol = 1e-04,
  verbose = TRUE)
}
\arguments{
\item{prior}{A list of priors for the parameters ab, gamma:
\itemize{
\item \code{sigma.inv.ab} a \eqn{(D+1)\times (D+1)} positive definite matrix corresponding to the inverse of the prior covariance matrix for \eqn{(\alpha, \beta)}.
\item \code{mu.ab} a \eqn{(D+1)} vector corresponding to the prior mean of \eqn{(\alpha, \beta)}.
\item \code{sigma.inv.gamma} a \eqn{D \times D} definite matrix corresponding to the inverse of the prior covariance matrix for \eqn{\gamma}.
\item \code{mu.gamma} a \eqn{D} vector corresponding to the prior mean of \eqn{\gamma}.
}}

\item{data}{A list of data values:
\itemize{
\item \code{Y} an \eqn{N \times J} matrix of \eqn{\pm 1} or \code{NA}.
\item \code{N}, \code{J} the dimensions of \code{Y}.
\item \code{D} the dimensions of political conflict being modeled.
}}

\item{init}{A list of initialization values for \code{alpha}, \code{beta}, \code{gamma}.}

\item{max.iter}{The maximum number of iterations.}

\item{tol}{The algorithm stops after the current iteration and the previous iteration parameters have correlation 1 - \code{tol}.}

\item{verbose}{print useful warnings and updates on algorithm progress.}
}
\value{
A list of estimated parameters \code{alpha}, \code{beta}, \code{gamma}
}
\description{
The user should never call this function directly! Use the
MultiScale function instead. The dense (as opposed to sparse)
algorithm includes the latent variables for missing vote
observations when updating the global parameters alpha, beta,
gamma. It is less robust in the presence of many missing votes
}
