% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{probit_log_prob}
\alias{probit_log_prob}
\title{Loss function for Bayes MAP estimation}
\usage{
probit_log_prob(params, data, priors)
}
\arguments{
\item{data}{A list of data values:
\itemize{
\item \code{Y} an \eqn{N \times J} matrix of \eqn{\pm 1} or \code{NA}.
\item \code{N}, \code{J} the dimensions of \code{Y}.
\item \code{D} the dimensions of political conflict being modeled.
}}
}
\description{
Do not call directly! Used by fit_intercepts
}
