% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense_updates.R
\name{update_ab}
\alias{update_ab}
\title{Update Alpha, Beta Parameters (Dense)}
\usage{
update_ab(Z, gamma, prior, data)
}
\arguments{
\item{Z}{An \eqn{N \times J} matrix of latent variables.}

\item{gamma}{An \eqn{N \times D} matrix of actor ideal points.}

\item{prior}{A list of priors for the parameters ab, gamma:
\itemize{
\item \code{sigma.inv.ab} a \eqn{(D+1)\times (D+1)} positive definite matrix corresponding to the inverse of the prior covariance matrix for \eqn{(\alpha, \beta)}.
\item \code{mu.ab} a \eqn{(D+1)} vector corresponding to the prior mean of \eqn{(\alpha, \beta)}.
\item \code{sigma.inv.gamma} a \eqn{D \times D} definite matrix corresponding to the inverse of the prior covariance matrix for \eqn{\gamma}.
\item \code{mu.gamma} a \eqn{D} vector corresponding to the prior mean of \eqn{\gamma}.
}}

\item{data}{A list of data values:
\itemize{
\item \code{Y} an \eqn{N \times J} matrix of \eqn{\pm 1} or \code{NA}.
\item \code{N}, \code{J} the dimensions of \code{Y}.
\item \code{D} the dimensions of political conflict being modeled.
}}
}
\value{
An updated \eqn{J \times (D+1)} matrix of question
    intercepts and slopes
}
\description{
The user should never call this function directly!
}
