% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale.R
\name{multiscale}
\alias{multiscale}
\title{Fit a zero- or multi-dimensional spatial voting model.}
\usage{
multiscale(method = "sparse", ...)
}
\arguments{
\item{method}{One of \code{c("sparse", "dense", "intercepts")},
with \code{"sparse"} as the default. If \code{"intercepts"},
estimate an intercept-only model. Otherwise, estimate a model
with positive dimension, choosing \code{"dense"} only if there
is not a lot of missing data.}

\item{...}{Parameters to pass to estimation methods.}
}
\value{
A list of estimated parameter values for \code{alpha},
    \code{beta}, and \code{gamma} --- the latter two not included
    if \code{"intercepts"} is chosen
}
\description{
Wrapper function for \code{multiscale_sparse} or
\code{multiscalle_dense}.
}
\examples{
\dontrun{
data(s109) # from pscl package
data(s109)
Y <- s109$votes  ## WARNING: data must be +/- 1, or NA
Y[Y \%in\% 1:3] <- 1
Y[Y \%in\% 4:6] <- -1
Y[Y \%in\% c(0, 7:9)] <- NA
data <- list(Y = Y, N = dim(Y)[1], J = dim(Y)[2], D = 2)
prior <- make_priors(data)
init <- make_starts(data)
lout <- multiscale(method = "sparse", prior = prior, data = data, init = init)
var(lout$gamma)
}
}
